+++ src/Components/Promises/PromiseTable/index.js
@@ -11,7 +11,7 @@ import {
 import InCompleteIcon from 'material-ui/svg-icons/navigation/close';
 import CompleteIcon from 'material-ui/svg-icons/navigation/check';
 import { red500, green500 } from 'material-ui/styles/colors';
-import type { PromiseType } from '../../types';
+import type { PromiseType } from '../../../types';
 
 export type Props = {
   promises: Array<PromiseType>,
