+++ src/Components/Promises/PromiseTable/stories/PromiseTable.stories.js
@@ -3,7 +3,7 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { muiTheme } from 'storybook-addon-material-ui';
 import PromiseTable from '../index';
-import { promises } from '../../../utils/testFixtures';
+import { promises } from '../../../../utils/testFixtures';
 
 storiesOf('Promise Table', module)
   .addDecorator(muiTheme())
